+++ test/porousmediumflow/2p2c/sequential/test_multiphysics2p2cproblem.hh
@@ -45,7 +45,7 @@ class TestMultTwoPTwoCProblem;
 // Specify the properties
 namespace Properties
 {
-NEW_TYPE_TAG(TestMultTwoPTwoCProblem, INHERITS_FROM(DecoupledTwoPTwoC, Test2P2CSpatialParams));
+NEW_TYPE_TAG(TestMultTwoPTwoCProblem, INHERITS_FROM(SequentialTwoPTwoC, Test2P2CSpatialParams));
 
 SET_TYPE_PROP(TestMultTwoPTwoCProblem, CellData, Dumux::CellData2P2CMultiPhysics<TypeTag>);